Using shellexpand for environment variables expansion on strings
Replaced old own implementation of environment variables expansion in
favor of more robust shellexpand library

+++ src/commands/base/load.rs
@@ -1,29 +1,30 @@
 use crate::core::error::Error;
-use crate::core::paths::expand_path;
 use crate::core::repo::Repo;
 use std::os::unix::fs;
+use std::path::PathBuf;
 
 pub fn load(force: &bool) -> Result<(), Error> {
     let current_dir = std::env::current_dir()?;
     let repo = Repo::load_at(current_dir)?;
 
     for (name, path_str) in &repo.config.files {
-        let path = expand_path(path_str);
-        if path.exists() {
-            if *force {
-                std::fs::remove_file(&path).map_err(|e| {
-                    Error::Msg(format!(
-                        "Failed to remove existing file {}: {}",
-                        path.display(),
-                        e
-                    ))
-                })?;
-            } else {
+        let path = PathBuf::from(shellexpand::full(path_str)?.to_string());
+
+        if path.exists() && !*force {
+            if !*force {
                 return Err(Error::Msg(format!(
                     "{} already exists. Use --force to overwrite.",
                     path.display()
                 )));
             }
+
+            std::fs::remove_file(&path).map_err(|e| {
+                Error::Msg(format!(
+                    "Failed to remove existing file {}: {}",
+                    path.display(),
+                    e
+                ))
+            })?;
         }
 
         fs::symlink(repo.root().join(name), &path).map_err(|e| {
